consumer/requirements.txt
confluent-kafka==2.8.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0